PyCon TW, Golden Pyramid, Elementary+
https://py.checkio.org/en/mission/golden-pyramid/

Our Robo-Trio need to train for future journeys and treasure hunts.
Stephan has built a special flat model of a pyramid.
Now the robots can train for speed gold running.
They start at the top of the pyramid and must collect
gold in each room, choose to take the left or right path
and continue down to the next level.
To optimise their gold runs, Stephan need to know
the maximum amount of gold that can be collected in one run.

Consider a tuple of tuples in which the first tuple
has one integer and each consecutive tuple
has one more integer then the last.
Such a tuple of tuples would look like a triangle.
You should write a program that will help Stephan
find the highest possible sum on the most profitable
route down the pyramid.
All routes down the pyramid involve
stepping down and to the left or down and to the right.