PyCon TW, Network Attack, Elementary+
https://py.checkio.org/en/mission/network-attack/

Nicola regularly inspects the local networks for security issues.
He uses a smart and aggressive program
which takes control of computers on the network.
This program attacks all connected computers simultaneously,
then uses the captured computers for further attacks.
Nicola started the virus program in the first computer and
took note of the time it took to completely capture the network.
We can help him improve his process by modeling and improving his inspections.

We are given information about the connections in the network and
the security level for each computer.
Security level is the time (in minutes)
that is required for the virus to capture a machine.
Capture time is not related to the number
of infected computers attacking the machine.
Infection start from the 0th computer (which is already infected).
Connections in the network are undirected.
Security levels are not equal to zero (except infected).

Information about a network is represented as a matrix NxN size,
where N is a number of computers.
If ith computer connected with jth computer,
then matrix[i][j] == matrix[j][i] == 1, else 0.
Security levels are placed in the main matrix diagonal,
so matrix[i][i] is the security level for the ith computer.