PyCon TW, Digit Stack, Elementary+
https://py.checkio.org/en/mission/digit-stack/

In computer science, a stack is a particular kind of data
type or collection in which the principal operations in the
collection are the addition of an entity to the collection
(also known as push) and the removal of an entity
(also known as pop). The relation between the push and
pop operations is such that the stack is a Last-In-First-Out
(LIFO) data structure. In a LIFO data structure, the last
element added to the structure must be the first one to be
removed. Often a peek, or top operation is also implemented,
returning the value of the top element without removing it.

We will emulate the stack process with Python.
You are given a sequence of commands:
- "PUSH X" -- add X in the stack, where X is a digit.
- "POP" -- look and remove the top position.
If the stack is empty, then it returns 0 (zero) and does nothing.
- "PEEK" -- look at the top position.
If the stack is empty, then it returns 0 (zero).
The stack can only contain digits.

You should process all commands and sum all digits
which were taken from the stack ("PEEK" or "POP").
Initial value of the sum is 0 (zero).