GITHUB, Open Labyrinth, Moderate
https://py.checkio.org/en/mission/open-labyrinth/

The labyrinth has no walls, but bushes surround the path on each side.
If a players move into a bush, they lose.
The labyrinth is presented as a matrix (a list of lists):
1 is a bush and 0 is part of the path.
The labyrinth's size is 12 x 12 and the outer cells are also bushes.
Players start at cell (1,1). The exit is at cell (10,10).
You need to find a route through the labyrinth.
Players can move in only four directions:
South (down [1,0]), North (up [-1,0]),
East (right [0,1]), West (left [0, -1]).
The route is described as a string consisting of different characters:
"S"=South, "N"=North, "E"=East, and "W"=West.

Input: A labyrinth map as a list of lists with 1 and 0.

Output: The route as a string that contains "W", "E", "N" and "S".